Support ingress annotations
To maintain backwards compatibility, we define two annotations source.

The global one is $.Values.authgear.ingress.annotations.
It is applied to all ingress generated by this helm chart.
The originally hard-coded value is copied to here.

The specific one is .ingress.annotations.
It is applied only to the specific ingress.